ProbniMarko/requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
duckduckgo-search>=4.1.1
langdetect>=1.0.9
deep-translator>=1.11.4
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
requests>=2.25.1
logfire>=0.5.0
python-json-logger>=2.0.0
html5lib>=1.1  # Better HTML parsing support
soupsieve>=2.5  # Required by beautifulsoup4
python-logging-loki>=0.3.1  # For better logging
chardet>=5.2.0  # For character encoding detection
asyncio>=3.4.3  # For async/await support